+++ src/Homeworks/homework_1/hw_1.cs
@@ -27,13 +27,34 @@ namespace BW
             }   
             return Result; 
         }
+        public static string[] Reborn (string? Str1) {
+            string[] AddString = new string[Str1.Length];
+            for (int j = 0; j < Str1.Length; j++) {
+                for (int i = 0; i < Str1.Length; i++) {
+                    AddString[i] = Str1.Substring(i,1) + AddString[i];
+                }
+                Array.Sort(AddString);
+            }
+            return AddString;
+        }
         static void Main(string[] args) 
         { 
-            System.Console.Write("Input the string: ");
-            string? InputString = Console.ReadLine();
-            string[] Strings2 = Transpositions(InputString);
-            Array.Sort(Strings2);
-            Console.WriteLine("B-W transform: {0}, {1}", LastColumn(Strings2), Array.BinarySearch(Strings2, InputString)+1);
+            System.Console.WriteLine("If you want a direct conversation, input 1, otherwise 2");
+            string? chose = Console.ReadLine();
+            if (chose == "1") {
+                System.Console.Write("Input the string: ");
+                string? InputString = Console.ReadLine();
+                string[] Strings2 = Transpositions(InputString);
+                Array.Sort(Strings2);
+                Console.WriteLine("B-W transform: {0}, {1}", LastColumn(Strings2), Array.BinarySearch(Strings2, InputString)+1);
+            }
+            else if (chose == "2") {
+                System.Console.Write("Input the tranform string: ");
+                string? InputString = Console.ReadLine();
+                System.Console.Write("Input the end of the string number: ");
+                int n = Convert.ToInt32(Console.ReadLine());
+                Console.WriteLine("Original string: {0}", Reborn(InputString)[n-1]);
+            }
             Console.ReadKey(); 
         }
     }
